--- src/components/pages/FirstPageComponent.js
@@ -1,16 +1,0 @@
-import React from 'react'
-import './../../App.css'
-import Header from './../Header'
-import Footer from './../Footer'
-
-function FirstPageComponent() {
-  return (
-    <div className="app-wrapper">
-      <Header />
-      <hr />
-      <Footer />
-    </div>
-  )
-}
-
-export default FirstPageComponent
